[P4ADEV-1827] moved auth logic to a separate service
